requirements.txt
redis
black
rich
python-dotenv
python-twitter-v2